warning in demux.cpp fixed
git-svn-id: http://svn.fastmq.org/svn/zeromq/branches/v0.3.3@622 7451a71c-b6c8-49cc-8254-141fc5361860

+++ zmq/demux.cpp
@@ -153,7 +153,7 @@ void zmq::demux_t::destroy_pipe (pipe_t *pipe_)
     if (it != pipes.end ()) {
         pipe_->write_delimiter ();
         pipes.erase (it);
-        if (current >= pipes.size ())
+        if (current >= (int) pipes.size ())
             current = 0;
     }
 }
